Classes for individ funcs
In an effort to clean up the code for when I eventually implement it into my larger project I am creating classes for each function. Not sure just yet how it will look when it is finished but saving it anyway. 

It may end up not being used. who knows?